Hotfix for CreateSectionLabels with duplicates
If there are domain objects with the same name in a single document,
Sphinx only creates an anchor tag for the first one, so the 'ids' of the
subsequent ones will be empty. We only need to create a label for the
first one, as the others aren't linkable anyway.

Fixes #142.